fix gqlgen tests

+++ contrib/99designs/gqlgen/appsec_test.go
@@ -106,7 +106,7 @@ func TestAppSec(t *testing.T) {
 
 				// The last finished span (which is GraphQL entry) should have the "_dd.appsec.enabled" tag.
 				span := spans[len(spans)-1]
-				require.Equal(t, 1, span.Tag("_dd.appsec.enabled"))
+				require.Equal(t, float64(1), span.Tag("_dd.appsec.enabled"))
 
 				type ddAppsecJSON struct {
 					Triggers []struct {
